Add folder structure for components and inputs
