+++ src/App.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
+import QuestionaryForm from './views/QuestionaryForm'
 import './styles/desktop.css'
 import './styles/globals.css'
 
@@ -8,6 +9,7 @@ const App = () => {
   return (
     <>
       <h1 className='title'>{t('getStarted')}</h1>
+      <QuestionaryForm />
     </>
   )
 }
