+++ src/components/Question.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Question = () => {
+  return (
+    <div className='question-container'>
+      <p data-cy='question'>Where can we reach you?</p>
+    </div>
+  )
+}
+
+export default Question
