+++ src/components/inputs/TypeField.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const TypeField = () => {
+  return (
+    <div data-cy='email-container'>
+      <input data-cy='email' name='email' type='email' placeholder='Email' />
+      <button data-cy='done-btn'>Done</button>
+    </div>
+  )
+}
+
+export default TypeField
